Moved test activity to test folder
